requirements.txt
aiohttp>=2.2.5
alphabet_detector
beautifulsoup4
bleach
cachetools
configparser
currencyconverter
emoji
feedparser
future
geopy
git+https://github.com/python-telegram-bot/ptbcontrib.git
gitpython
gpytranslate
hurry.filesize
intellivoid-SPB
markdown2
multicolorcaptcha
nltk
pillow>=4.2.0
psutil
psycopg2-binary
pyYAML>=5.1.2
pyrate-limiter
pyrogram
python-telegram-bot==13.5
requests
rich
spamwatch
speedtest-cli
spongemock
sqlalchemy==1.4.11
telethon==1.16.4
tgcrypto
tswift
ujson
wikipedia
zalgo-text
